#22 [REFACTOR] Domain 모듈 패키지 구조 변경
