+++ presentation/src/main/res/layout/item_team_member.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="member"
-            type="com.tgyuu.domain.team.Member" />
+            type="com.tgyuu.domain.team.entity.Member" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
